+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to Haynoi are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.3.10] - 2026-09-01
 
 ### Fixed
 - **When our servers have a problem, Haynoi now says so** — instead of telling